pyproject.toml
[project]
name = "ragprod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Adrià Cabello", email = "adrian.cabello@alumni.esci.upf.edu" }
]
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.47.0",
    "chromadb>=1.1.0",
    "editables>=0.5",
    "fastmcp>=2.12.3",
    "hatchling>=1.27.0",
    "ipykernel>=6.30.1",
    "langfuse>=3.5.0",
    "openai>=1.108.1",
    "opentelemetry-sdk>=1.37.0",
    "pydantic>=2.11.7",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "sentence-transformers>=5.1.0",
    "together>=1.5.25",
]

# Specific versions of torch, torchvision, and torchaudio are installed based on the extra feature uv sync --extra cu124
[project.optional-dependencies]
cpu = ["torch==2.4.1", "torchvision==0.19.1", "torchaudio==2.4.1"]
cu124 = ["torch==2.4.1", "torchvision==0.19.1", "torchaudio==2.4.1"]

[tool.uv]
# The cpu and cu124 extras are mutually exclusive, you can't use both at the same time
conflicts = [[{ extra = "cpu" }, { extra = "cu124" }]]

# The package is only installed on Linux systems
# environments = ["sys_platform == 'linux' and os_name == 'posix'"]

[tool.uv.sources]
# The torch, torchvision, and torchaudio packages are installed from the pytorch-cpu and pytorch-cu124 indexes
torch = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torchvision = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]
torchaudio = [
  { index = "pytorch-cpu", extra = "cpu" },
  { index = "pytorch-cu124", extra = "cu124" },
]

[[tool.uv.index]]
# The pytorch-cpu index is used to install the torch, torchvision, and torchaudio packages
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[project.scripts]
ragprod = "ragprod:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
